Move the new Futures to Edge namespace
removing the ConcurrentNext space
