+++ lib/concurrent-edge.rb
@@ -2,5 +2,6 @@ require 'concurrent'
 
 require 'concurrent/actor'
 require 'concurrent/channel'
+require 'concurrent/edge/future'
