+++ lib/concurrent/file_map.rb
@@ -6,7 +6,7 @@ module Concurrent
                        'lib/concurrent/actor/**/*.rb',
                        'lib/concurrent/channel.rb',
                        'lib/concurrent/channel/**/*.rb',
-                       'lib/concurrent/next.rb'] & git_files
+                       'lib/concurrent/edge/**/*.rb'] & git_files
   core_lib_files = all_lib_files - edge_lib_files
 
   FILE_MAP = {
