+++ spec/spec_helper.rb
@@ -21,8 +21,7 @@ end
 
 $VERBOSE = nil # suppress our deprecation warnings
 require 'concurrent'
-require 'concurrent/actor'
-require 'concurrent/channel'
+require 'concurrent-edge'
 
 logger                          = Logger.new($stderr)
 logger.level                    = Logger::WARN
